Move Pro Ruby code to MIT license
Moved lib/react_on_rails/pro/ code to MIT-licensed locations:
- lib/react_on_rails/pro/helper.rb → lib/react_on_rails/pro_helper.rb
- lib/react_on_rails/pro/utils.rb → lib/react_on_rails/pro_utils.rb

Changes:
- Removed Pro license headers from moved files
- Changed namespace from ReactOnRails::Pro:: to ReactOnRails::
- Removed Pro warning badge functionality (no longer needed)
- Deleted lib/react_on_rails/pro/ directory entirely
- Updated LICENSE.md to reflect all lib/react_on_rails/ is MIT
- Updated documentation references

Benefits:
- Ruby signaling code now in MIT gem for marketing visibility
- Actual functionality still requires Pro npm package
- Clean license boundaries maintained
- Users can't bypass Pro features by editing Ruby code
