+++ LICENSE.md
@@ -11,7 +11,7 @@ This repository contains code under two different licenses:
 
 The following directories and all their contents are licensed under the **MIT License** (see full text below):
 
-- `lib/react_on_rails/` (excluding `lib/react_on_rails/pro/`)
+- `lib/react_on_rails/` (entire directory)
 - `packages/react-on-rails/` (entire package)
 - All other directories in this repository not explicitly listed as Pro-licensed
 
@@ -19,7 +19,6 @@ The following directories and all their contents are licensed under the **MIT Li
 
 The following directories and all their contents are licensed under the **React on Rails Pro License**:
 
-- `lib/react_on_rails/pro/`
 - `packages/react-on-rails-pro/` (entire package)
 - `react_on_rails_pro/` (entire directory)
