--- lib/react_on_rails/pro/NOTICE
@@ -1,21 +1,0 @@
-# React on Rails Pro License
-
-The files in this directory and its subdirectories are licensed under the **React on Rails Pro** license, which is separate from the MIT license that covers the core React on Rails functionality.
-
-## License Terms
-
-These files are proprietary software and are **NOT** covered by the MIT license found in the root LICENSE.md file. Usage requires a valid React on Rails Pro license.
-
-## Distribution
-
-Files in this directory will be **omitted** from future distributions of the open source React on Rails Ruby gem. They are exclusively available to React on Rails Pro licensees.
-
-## License Reference
-
-For the complete React on Rails Pro license terms, see: `REACT-ON-RAILS-PRO-LICENSE.md` in the root directory of this repository.
-
-## More Information
-
-For React on Rails Pro licensing information and to obtain a license, please visit:
-- [React on Rails Pro](https://www.shakacode.com/react-on-rails-pro/)
-- Contact: [react_on_rails@shakacode.com](mailto:react_on_rails@shakacode.com)
